Add runnable scripted and remove bloop dep
Scripted can now be run like any kind of other application. As there was
some problem in the community build with depending on sbt-complete, that
dependency has been sacrificed and instead we handle the input.

The input of the new runnable scripted is a list of scripted filters
followed by an optional 'true/false' argument to enable/disable log
buffering.

The most notorious improvement caused by this change is that now we
don't need to resolve and compile the bridge, but instead we just get
the output of classes directory of every bridge, put it in a jar and
classload it. This speeds up test execution.

After this change, the bloop scripted module is no longer necessary as
bloop can run the application in the `zincScripted-test` target.

All tests run in 97s after this change.
