+++ internal/zinc-scripted/src/test/scala/sbt/inc/ScriptedBridgeProvider.scala
@@ -4,20 +4,20 @@ import java.io.File
 
 import sbt.internal.inc.classpath.ClasspathUtilities
 import xsbti.Logger
-import xsbti.compile.{ CompilerBridgeProvider, ScalaInstance }
-import bloop.config.Config
+import xsbti.compile.CompilerBridgeProvider
 
-final class BloopBridgeProvider(
-    defaultBridge: Config.Project,
-    bridges: List[Config.Project],
+case class ScalaBridge(version: String, jars: Seq[File], classesDir: File)
+
+final class ScriptedBridgeProvider(
+    bridges: List[ScalaBridge],
     tempDir: File
 ) extends CompilerBridgeProvider {
-  import BloopBridgeProvider.XScala
 
-  /**
-   * Get the location of the compiled Scala compiler bridge for a concrete Scala version.
-   */
+  import xsbti.compile.ScalaInstance
+
+  /** Get the location of the compiled Scala compiler bridge for a concrete Scala version. */
   override def fetchCompiledBridge(instance: ScalaInstance, logger: Logger): File = {
+
     /** Generate jar from compilation dirs, the resources and a target name. */
     def generateJar(outputDir: File, targetJar: File) = {
       import sbt.io.IO._
@@ -28,17 +28,18 @@ final class BloopBridgeProvider(
       targetJar
     }
 
-    val scalaVersion = instance.version()
-    bridges.find(_.getScala.version == scalaVersion) match {
-      case None => sys.error(s"Missing $scalaVersion in ${bridges.map(_.project.name).mkString}")
+    val scalaVersion = instance.version
+    bridges.find(_.version == scalaVersion) match {
+      case None =>
+        sys.error(s"Missing $scalaVersion in supported versions ${bridges.map(_.version).mkString}")
       case Some(bridge) =>
         val targetJar: File = {
           import sbt.internal.inc.ZincComponentCompiler.{ binSeparator, javaClassVersion }
-          val id = s"${bridge.name}$binSeparator${bridge.getScala.version}__$javaClassVersion.jar"
+          val id = s"scriptedCompilerBridge$binSeparator${bridge.version}__$javaClassVersion.jar"
           tempDir.toPath.resolve(id).toFile
         }
 
-        generateJar(bridge.classesDir.toFile, targetJar)
+        generateJar(bridge.classesDir, targetJar)
     }
   }
 
@@ -52,13 +53,12 @@ final class BloopBridgeProvider(
    * @see CompilerBridgeProvider#fetchCompiledBridge
    */
   override def fetchScalaInstance(scalaVersion: String, logger: Logger): ScalaInstance = {
-
-    bridges.find(_.getScala.version == scalaVersion) match {
-      case None => sys.error(s"Missing $scalaVersion in ${bridges.map(_.project.name).mkString}")
+    bridges.find(_.version == scalaVersion) match {
+      case None =>
+        sys.error(s"Missing $scalaVersion in supported versions ${bridges.map(_.version).mkString}")
       case Some(bridge) =>
-        val scala = bridge.getScala
         import sbt.internal.inc.ScalaInstance
-        val jars = scala.jars.toArray.map(_.toFile)
+        val jars = bridge.jars.toArray
         assert(jars.forall(_.exists), "One or more jar(s) in the Scala instance do not exist.")
 
         val libraryJar = jars.filter(_.getName.contains("scala-library")).head
@@ -68,12 +68,5 @@ final class BloopBridgeProvider(
         val loader = ClasspathUtilities.toLoader(missingJars, libraryL)
         new ScalaInstance(scalaVersion, loader, libraryL, libraryJar, compilerJar, jars, None)
     }
-  }
-}
-
-object BloopBridgeProvider {
-  implicit class XScala(val project: Config.Project) extends AnyVal {
-    def getScala: Config.Scala =
-      project.scala.getOrElse(sys.error(s"No scala found in bridge ${project.name}"))
   }
 }
