+++ project/plugins.sbt
@@ -5,3 +5,4 @@ addSbtPlugin("de.heikoseeberger" % "sbt-header" % "3.0.2")
 addSbtPlugin("com.thesamet" % "sbt-protoc" % "0.99.12-rc5")
 libraryDependencies += "com.trueaccord.scalapb" %% "compilerplugin" % "0.6.0"
 addSbtPlugin("com.lightbend" % "sbt-whitesource" % "0.1.9")
+addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.9.0")
